requirements.txt
fastapi
uvicorn
python-telegram-bot
supabase
requests
beautifulsoup4